--- codegen/aws-sdk-codegen/src/main/kotlin/aws/sdk/kotlin/codegen/customization/sfn/SfnLongPollingIntegration.kt
@@ -1,40 +1,0 @@
-/*
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-package aws.sdk.kotlin.codegen.customization.sfn
-
-import aws.sdk.kotlin.codegen.sdkId
-import aws.smithy.kotlin.codegen.KotlinSettings
-import aws.smithy.kotlin.codegen.core.KotlinWriter
-import aws.smithy.kotlin.codegen.core.RuntimeTypes
-import aws.smithy.kotlin.codegen.integration.KotlinIntegration
-import aws.smithy.kotlin.codegen.rendering.protocol.ProtocolGenerator
-import aws.smithy.kotlin.codegen.rendering.protocol.ProtocolMiddleware
-import software.amazon.smithy.model.Model
-import software.amazon.smithy.model.shapes.OperationShape
-import software.amazon.smithy.model.shapes.ServiceShape
-
-private val LONG_POLLING_OPERATIONS = setOf("GetActivityTask")
-
-class SfnLongPollingIntegration : KotlinIntegration {
-    override fun enabledForService(model: Model, settings: KotlinSettings): Boolean = model.expectShape(settings.service, ServiceShape::class.java).sdkId.lowercase() == "sfn"
-
-    override fun customizeMiddleware(
-        ctx: ProtocolGenerator.GenerationContext,
-        resolved: List<ProtocolMiddleware>,
-    ): List<ProtocolMiddleware> = resolved + LongPollingMiddleware(LONG_POLLING_OPERATIONS)
-}
-
-private class LongPollingMiddleware(private val operationNames: Set<String>) : ProtocolMiddleware {
-    override val name: String = "LongPolling"
-
-    override fun isEnabledFor(ctx: ProtocolGenerator.GenerationContext, op: OperationShape): Boolean = op.id.name in operationNames
-
-    override fun render(ctx: ProtocolGenerator.GenerationContext, op: OperationShape, writer: KotlinWriter) {
-        writer.write(
-            "op.context[#T.LongPolling] = true",
-            RuntimeTypes.HttpClient.Operation.HttpOperationContext,
-        )
-    }
-}
